requirements.txt
flask>=3.0,<4.0
pymongo>=4.6,<5.0
redis>=5.0,<6.0
python-dotenv>=1.0,<2.0
bson>=0.5.10
